Fix RuboCop settings
RuboCop was failing on Travis because it picked up TargetRubyVersion
from the Rainbow gem's .rubocop.yml. It was picking it up, because we
didn't explicitly exclude files from `vendor/bundle`. That directory is
excluded by default, but we overrode the default without merging the
parent configuration.

Therefore, the fix is to merge our Exclude into the parent (default)
config.